gtest genesis parsing (nodes/rust/duniter-v2s!176)
* fix clippy

* fix after rebase

* WIP example specs

* fix after rebase

* add reproducibility

* doc add missing protobuf compiler

* remove ref to genesis timestamp

* fix broken next identity index genesis

* our pallets are not coming from the framework

* fix first_ud everywhere

* remove files unwanted on main branch

* also apply renaming to gdev

* rename first_ud_value and add comments

(to put in value the absence of first_ud_time)

* use index provided in GenesisIdty

* add mold to documentation

* change gtest existential deposit

and adapt genesis parsing to real-world data

* embed raw chainspecs only when enabled

* add comments

+++ node/src/chain_spec/gdev.rs
@@ -339,7 +339,7 @@ fn gen_genesis_for_local_chain(
     assert!(initial_smiths_len <= initial_identities_len);
     assert!(initial_authorities_len <= initial_smiths_len);
 
-    let babe_epoch_duration = get_env("DUNITER_BABE_EPOCH_DURATION", 30);
+    let babe_epoch_duration = get_env("DUNITER_BABE_EPOCH_DURATION", 30) as u64;
     let cert_validity_period = get_env("DUNITER_CERT_VALIDITY_PERIOD", 1_000);
     let membership_period = get_env("DUNITER_MEMBERSHIP_PERIOD", 1_000);
     let smith_cert_validity_period = get_env("DUNITER_SMITH_CERT_VALIDITY_PERIOD", 1_000);
@@ -505,7 +505,7 @@ fn gen_genesis_for_benchmark_chain(
     assert!(initial_smiths_len <= initial_identities_len);
     assert!(initial_authorities_len <= initial_smiths_len);
 
-    let babe_epoch_duration = get_env("DUNITER_BABE_EPOCH_DURATION", 30);
+    let babe_epoch_duration = get_env("DUNITER_BABE_EPOCH_DURATION", 30) as u64;
     let cert_validity_period = get_env("DUNITER_CERT_VALIDITY_PERIOD", 1_000);
     let membership_period = get_env("DUNITER_MEMBERSHIP_PERIOD", 1_000);
     let smith_cert_validity_period = get_env("DUNITER_SMITH_CERT_VALIDITY_PERIOD", 1_000);
